Mucfgexpand: Rework VT-d root/IR table handling
Also expand VT-d root and interrupt remapping tables in case of disabled
IOMMU support. This is in preparation of making such images work with
Tau0 (Tau0 enforces an IOMMU since non-IOMMU systems are considered a
development/Bochs feature).

If the IOMMU is disabled, add 0-filled memory regions for VT-d root and
IR tables. The fill content is required to guarantee that the allocator
places the regions in lower memory. Since the regions are typed via
clearing in the Tau0 command stream, they will be part of the image and
therefore have an effect on image size.