Rework scheduling plan handling
Do not care about scheduling plan generation if a <scheduling> node is
already present.

The new static subdirectory contains pre-defined plans for
SYSTEM-PLATFORM.xml or SYSTEM.xml policies, whereas the former take
precedence over the latter.

If no static plan exists, it it is assumed that a plan needs to be
generated, for which a per-system scheduling config must be present.
Fail if no such config is found.

To test the feature, factor out the static plans of integration_tests_kt
and mirage-solo5 policies.