Mucfgexpand: Fix typo in library resource expander
The typo effectively lead to <library> elements not being treated as
list tags when merging library into component specifications. However,
practically this had no effect because library dependency resolution is
implemented recursively. This means that all nested depends <library>
nodes are resolved prior to merging and thus it is not possible for them
to be mishandled. Correct the typo instead of dropping library as a list
tag to be safe from future code changes.