Add Unmask IRQ logical event validator
The validator checks that logical names of events with unmask_irq event
action conform to the naming scheme: it must start with 'unmask_irq' and
end with "_$VecNr" where $VecNr is the numeric value of the
corresponding logical device IRQ vector.

Furthermore, the validator checks that the unmask action number matches
the corresponding physical IRQ number.