contrib: Add GRUB2 (2.04) with CSL patch
The Command Stream Loader (CSL) GRUB2 module enables the processing of
TLV-encoded commands. The current protocol version specifies the
following commands:

- CMD_CHECK_VERSION
- CMD_META_DATA (ignored)
- CMD_WRITE
- CMD_FILL
- CMD_SET_ENTRY_POINT
- CMD_CHECK_CPUID

The CSL module solves the problem of loading big images on targets with
memory-holes and allows to perform additional (CPUID-based) checks on
the platform.

The existing 'multiboot' module is hacked to compare the CSL generated
memory image with the one generated via multiboot loader. This asserts
in-memory image equivalence for CI.