Use GRUB2 CSL to load Muen system image
Load and process the Tau0-generated TLV command stream with the GRUB2
CSL module.

The additional multiboot module line in grub.cfg is used to compare the
CSL in-memory image with the multiboot image. The boot aborts if a
binary difference exists. This is used to ensure that the CSL module
produces an identical image.