Add hardware config for QEMU/KVM with OVMF bios
Note that the names of the ethernet controllers are adjusted to match
the ordering in qemu-kvm.xml.

Rationale: ethernet_controller_1 is referenced without alias in
common_device_domains.xml and also in the pciconf_emulation_nic_physdev
variable. While it is possible to use aliases/classes throughout the
demo systems, adjusting the names seems less invasive.