emulate: Add support for EFI boot
If the MUEN_EFI environment variable is set, adjust QEMU_OPTS to start
QEMU with EFI bios.

Note that VNC instead of curses is used for EFI boot as EFI requires a
GFX console.  This enables the screen for OVMF and GRUB2 until EFI boot
services are turned off. After that, no VT output is visible because
Muen currently lacks an EFI-aware console.

Also, do not call the multiboot command in grub.cfg when running with
EFI (comparing in-memory images does not work on EFI because of memory
holes -> allocated memory is not consecutive).