Travis: Only initialize Linux submodule with --depth=1
Otherwise git is only able to fetch advertised refs (i.e. branches or
tags). This is problematic e.g. if a devel branch of a submodule
contains new commits, but the submodule is still pinned to an older
commit.

Some sources advertise to increase the --depth value to solve this
problem, but this did not help.

Allowing unadvertised objects on the server side is no option as well,
since it may be computationally expensive to calculate reachability of
an object (see documentation about uploadpack.allowReachableSHA1InWant).