Restrict hypercalls to subject privilege level 0
This has the effect that source events can only be triggered from
kernel-space. It is primarily relevant for VM subjects.

It should be noted that the system policy is and always has been
correctly enforced by the Muen Kernel. It is a measure catering to VM
subjects, in order to restrict which software parts inside the subject
can trigger source events.