Add -mindirect-branch=thunk-extern to compiler switches
Part of GCC retpoline mitigation for Spectre v2. Instructs the compiler
to replace indirect calls and jumps with call and return thunk,
see [1]. thunk-extern tells the compiler that the call and return thunk
(__x86_indirect_thunk_rax) is provided in a separate object file.

== Note1

There are currently no indirect jumps/calls in both the SK and the
native SPARK/Ada components.

== Note2

This purely serves as a way to detect unwanted indirect jumps/calls, as
we do not provide an actual __x86_indirect_thunk_rax implementation.

Linking objects containing an indirect call leads to the following
error:

test.c: undefined reference to `__x86_indirect_thunk_rax'

== Note3

We do not enable -mfunction-return=thunk as this is way to expensive.

[1] - https://gcc.gnu.org/onlinedocs/gcc/x86-Options.html