Make file-backed regions reloadable if necessary
A component is considered reloadable if it declares a dependency to
libmuinit in the XML specification. In that case, each file-backed
memory region is made non-writable and mapped high. Then, a writable
memory region of the same size and with the same content hash is mapped
at the original virtual address.