Drop debug messages
This change makes SL stateless and thus removes the need for a data
section. On top, the log messages no longer add much additional
information since the state can be determined via the corresponding
status page. Furthermore, the SM subject emits messages when handing
over execution to SL and when it returns.