Add isolation tests monitor component
The isolation test monitor works in tandem with the isolation tests
subject. Whenever the isolation test subject traps to the monitor, it
saves the state of the ITS subject to the result/trapped state region,
advances the RIP to the next instruction and hands execution back to the
isolation test runner to resume test execution.