Policy: Make ahci_drv_enabled flag work
To make ahci_drv_disable=false in demo_system_vtd.xml work, the
muenblock client code in the example component must be disabled. Also,
all associated channels and the ahci_drv subject itself must not be
included in the system policy in this case.

Note: It is still required to manually remove the ahci_drv subject from
      the scheduling plan configuration atm.