Make muenblock client channel types suitable for aliasing
Create explicit subtypes with object size aspect to make the shared
memory channel type suitable for aliasing via address clause.

Currently, there is a limitation in Ada/SPARK, which does not allow the
specification of the memory channel size within the Muchannel generic,
since 'Size must be a static expression (see also ticket [T611-019]).

A language enhancement [1] is in discussion which will make this
workaround obsolete.

[1] - https://github.com/AdaCore/ada-spark-rfcs/issues/75