Drop Vector parameter from Setup_Fault_Interrupt
Since the value is hardcoded to SK.Constants.VTd_Fault_Vector there is
no need for the caller to pass it in as a parameter. This avoids
with'ing SK.Constants and having to disable the "unreferenced" warning
in the SK.VTd package body, since it is only used in debug.