Handle XSETBV VM-exit
This enables subjects to control their own XCR0 value and thus the
active FPU features.

Validate the given XCR0 value according to Intel SDM Vol. 1, "13.3
Enabling the XSAVE Feature Set and XSAVE-enabled Features" and inject
a #GP if a check fails. If the value is valid, set the corresponding
subject XCR0 value, increment the RIP and resume subject execution.