Set per-subject XCR0 value after FPU state restore
Do a full restore of the subject FPU state region and then set the
associated XCR0 value which might enable less extended CPU features in
the future.

To ensure that the full FPU state is saved, we set the XCR0
unconditionally to all active FPU features prior to executing xsave.

Note: This might be overly cautious since according to the xrstor
      instruction specification in Intel SDM Vol. 2, it is supposed to
      default-initialize parts of the FPU state which are
      inactive/unused. Microbenchmarking this variant did not show
      decreased VM-Exit handler cycle count so there seems to be no
      adverse effect on doing an explicit full save/restore.