Move crash audit region
This is required because on the T480s, a yet to be identified entity
(firmware?) overwrites the first 64-bit word with the value 2^32, which
is the physical address of the crash audit region itself
(16#1_0000_0000#).

Boot loaders (ipxe and grub) can be ruled out, as they cannot reach this
memory (they have no support for PAE in 32bit-mode).

Therefore, use a more "random" location for all targets.