Mucfgvalidate: Unregister Shared_Device_Same_PCI_Element check
This validator is too strict for the generic case where PCI devices are
assigned to multiple subjects. It was intended to cover the Device
Manager/PCI Config Space emulation use case. However, a misconfiguration
of that scenario leads to an error message during Linux boot, which is
precise enough to deduce the source of the issue (see also T615).