Set type of initramfs loader src regions to 'subject'
The source regions are regular subject memory with contents that are
copied to the associated target regions. The target region needs to be of
the original memory type since it fulfills that function, the source
region however does not. This avoids false positives with the initramfs
consecutiveness expander since there is no requirement for the source
regions to be consecutive *in the loader* address space.

Note that adjusting the validator is not possible because in policy
format B, the case where consecutiveness matters (i.e. Linux subjects)
and where it does not matter (i.e. Loaders) can not be distinguished.