Add scheduling partitions to desktop demo system policies
Add scheduling partitions and scheduling groups to the desktop demo
system policies. Adapt the smpify.py script to add scheduling partitions
and group definitions for additional CPU cores.
The specification is almost equivalent to the existing/implicitly defined
scheduling groups: the scheduling group IDs are a bit reordered for time
and dbgserver to simplify the addition of scheduling partitions in the
smpify.py script.