Add Subject_[Sleep|Yield] to Event_Action_Kind type
Since the type is propagated via generated specs, the affected toolchain
tests have been adapted.

On the kernel side, the two source actions are implemented as null
actions for now. The kernel change is necessary as part of this commit
since otherwise the compiler would rightfully complain, that the case
statement does not cover all event action kinds.