Add Scheduler.Global_Group_Activity_Indicator
The new variable will be used to track the activity state of scheduling
groups for each scheduling partition. It is globally accessible since
CPUs handling events with a different target CPU may atomically set the
activity flag of the scheduling group of the event target subject. Cause
it is volatile, it is modeled as a separate, external abstract state to
avoid making the whole existing scheduler state volatile.