Add Running field to subject state
The record field will be used by the kernel to keep track of whether a
subject is running or sleeping. By default running is set to True. It is
explicitly not configurable via the policy to avoid implementation
complexity and exclude the possibility that the system may get stuck
with a badly timed subject reset event.

Note: While the new field is accessible to subject monitors, the kernel
      explicitly gives no guarantee regarding the data format in order
      to avoid codifying an ABI.

Note: Explicit 7-bit padding is added as otherwise GNATprove (correctly)
      complains that the subject state type is not suitable for aliasing
      via Address clause. Otherwise these 7 bits would be undefined.