Update scheduling partition on minor frame change
Whenever a new minor frame is entered, the now active scheduling
partition is updated by looking up the next active scheduling group. If
there is an active group, it is chosen as the new active group of the
partition and execution will continue with the active subject of that
new scheduling group. If no group is active, then scheduling partition
transitions into the sleep state.

The scheduling group activity will be reevaluated at the next minor
frame start.