Implement scheduling group deactivation
Clear the global activity indicator of the scheduling group identified
by partition ID and scheduling group index. Also insert the timed event
of the active subject of the scheduling group into the sorted timer list
of the scheduling partition if it is not already in the list. An
inactive scheduling group might be deactivated if the corresponding
activity indicator was set due to an event from a different core.