Implement rescheduling of partitions by subjects
Subjects can use the yield and sleep source event actions to trigger a
rescheduling of their partition.
For the remainder of the current minor frame, the next active scheduling
group in the partition will be switched to or if all are inactive, the
scheduling partition is, as a whole, put into sleep mode.