Indicate activity when triggering event
Toggle activity flag of target subject/scheduling group when a target
event is newly marked pending. If the source of the event is on a
different CPU we always mark the group pending. Otherwise, we check if
the target subject is actually the active subject of a scheduling group
and only activate if it is.

Note: Marking a group as pending is not an issue since it will only lead
      to a potentially unnecessary check if a scheduling group is in fact
      active but never to the activation of a scheduling group that is
      actually inactive.