Evaluate timed events of inactive scheduling groups
Check the scheduling partition timer list as part of performing a
scheduling decision. The timer list contains all inactive scheduling
groups of a partition. Traverse the list to activate all groups whose
timer has expired and update the list as necessary, removing now active
groups and adjusting the head of the timer list.