Example: Make Foo interface asynchronous
Instead of switch events use asynchronous notifications. Adjust example
component code to handle incoming requests. With the upcoming use of
sleep/yield operations, this will nicely illustrate how to implement
asynchronous interfaces in components without busy looping.

It is also a preparatory step to be able to move the example subject to
a separate scheduling group and test scheduling partitions with multiple
groups.