Update Muenfs submodule to a554aa0109d8
Contains example test program that illustrates how event-driven, shared
memory channel based interfaces can be implemented and in particular how
to use poll in combination with muenfs/channel events. It will be used
to exercise this functionality in CI.

Also includes a fix for a race condition wrt. registering the IRQ
handler.