Recheck activity when rescheduling subject
After deactivation, check if the subject has become active in the
meantime as subjects on other cores may send events at any time. If it
has become active, set the global indicator activity of the scheduling
group again. This is important since the activity indicator being set is
a precondition to the evaluation whether a group is actually active or
not (see Find_Next_Active_Scheduling_Group).