Set active subject to running on partition reschedule
When a partition is rescheduled and a new scheduling group becomes
active, set the running flag of its active subject. This makes sure
that the now scheduled subject does not go back to sleep after being
executed for one minor frame.

Consider the following case: a subject yields or sleeps and the subject
of the next group was previously sleeping and has now just become
active, e.g. due to a newly pending event. If the Running flag is not
set, then it will no longer be categorized as active once it has run for
a minor frame and has no pending events or interrupts (Is_Active will
return false. As a consequence the group will go back to sleep even
though that action was not explicitly requested.