GHA: Only save caches if on devel branch
A cache is only accessible if its associated branch is the branch which
triggered the GHA workflow, or the default branch of the repository
(i.e.  devel).

Therefore use the cache restore/save actions to only store the cache if
the workflow runs in the devel branch context. This way, feature
branches can use this cache and won't evict it.

Note: 'Cache save failed' warnings will be displayed if the devel branch
changes, see issue https://github.com/actions/cache/issue/1139. These
warnings are expected and can be ignored.