Only process events & ints when partition is active
Do not handle target events and inject pending interrupts in the case,
when the current partition is sleeping. Wake up of partitions only
happens whenever it is rescheduled due to a minor/major frame switch.
While a partition is inactive due to the sleep state, pending events and
interrupts must not be processed. Otherwise, a partition might not wake up
properly.