Set subject running if handling pending interrupt/event
When injecting an interrupt or handling a target event of a subject that
is sleeping, set the running flag. This ensures that the subject will
not go to sleep after the pending interrupt/event has been cleared while
at the same time enabling immediate wakeup for partitions which are
asleep.

This commit reverts d971d126154e since it solves the issue in a more
efficient way.