Kernel: Optimize Is_Active by prioritizing Running flag
In the regular case, the Running flag of a subject is set to True. Thus
disentangle the various conditions by forcing the compiler to check the
conditions in a serial manner and check the Subject Running flag first.
This avoids a lot of work in the regular case.

Note: The compiler would generate optimal code with a concatenated
      "or else" condition. However, SPARK does not allow calling
      volatile functions in this context. Thus the condition is
      formulated in the unfolded manner.