SM: Greatly simplify CPU_Values package
And therefore speedup CPUID emulation, which is triggered quite
frequently by Linux/glibc etc.

Instead of generating arrays of supported CPUID leaf/subleafs and MSRs,
generate constants instead. This avoids looping over the arrays to find
the supported entries.

The code for both CPUID and MSRs expects specific CPUID leafs/MSRs to be
present anyway, for which a constant must be available in the generated
CPU_Values.Target package. This works since we only support basic CPUID
leafs and MSRs, which are present on all supported hardware targets.