Extract UART character send count to separate package
This allows users to provide their own instance of the UART FIFO send
count variable which is necessary because the kernel has to place it in
the global data section shared by all cores. Without this change, each
core has their own, local copy which can lead to garbled UART output
when all CPU cores write to it simultaneously.