Libmutime: Introduce Multiply_Divide procedure
The procedure multiplies a given value with a multiplier and then
divides it by the specified divisor.

The procedure supports 128 bit arithmetic by using the processor's MUL
and DIV instructions. It basically calculates the following efficiently:

Quotient := Unsigned_64'Mod (Unsigned_128 (Value) * Unsigned_128
  (Multiplier) / Unsigned_128 (Divisor))

The processor will raise an exception if the quotient is >= 2 ** 64. A
precondition ensures that this does not happen.