Libmutime: Use Multiply_Divide in Get_Current_Time
Commit 4c5159775 switched from MHz to Hz in offset calculation to avoid
time drift over time. Unfortunately, there is still a problem with this
approach: it introduces a time quantization of 1s because the offset is
divided before the multiplication.

The solution is to use 128 bit arithmetic via the new Multiply_Divide
procedure, in order to perform the multiplication before the division,
without loss of precision.