Save/Restore FPU state in the interrupt handler
Explicitly save FPU state at beginning of Ada/SPARK interrupt handler
and restore it at the end using XSAVE/XRSTOR.  This enables the
invocation of subprograms which make use of the FPU, i.e. optimized Put
operations which use XMM registers for faster string copying. Otherwise
the FPU state of interrupted code is clobbered.

This is to illustrate how to implement interrupt handlers that may use
the FPU. An alternative is to disallow FPU use in the interrupt handler
by using the attribute `no_caller_saved_registers` in combination with
the `-mgeneral-regs-only` compiler flag. To apply the attribute to
Ada/SPARK subprograms the following snippet can be used:

  pragma Machine_Attribute
    (Entity         => $Subprogram_Name,
     Attribute_Name => "no_caller_saved_registers");