Add SK.CPU.Wait_For_Interrupt procedure
The subprogram halts instruction execution until the occurrence of the
next Interrupt. This is achieved by putting sti and hlt in the same code
block. The instruction sequence needs to be part of the same assembler
code block since sti only blocks interrupts for the next instruction
immediately following it. Otherwise, if an interrupt occurs between sti
and the eventual hlt, the code might stay halted until the next
interrupt occurs.