Policy: Rewire Linux 1 reboot to Slot 1 reset request
Instead of just reseting itself, the necessary actions of a Linux reboot
include resetting the SM also, i.e. Linux needs to request a reset of
the slot it belongs to. Map the Linux reboot event to the appropriate
slot reset request event and hey presto, Linux reboot works again!