Add --z noexecstack linker option
Recent version of the GNU linker ld from binutils 2.39+ will emit the
following warning:

  ld: warning: stub.o: missing .note.GNU-stack section implies executable stack
  ld: NOTE: This behaviour is deprecated and will be removed in a future version of the linker

While the general solution would be to add a .note.GNU-stack section to
all of our assembly files (*.S) [1], whether the stack of a component is
executable or not is specified in the component XML spec/policy. So in
our case, these specific ELF sections have no effect and it is simpler
to use the linker switch instead. Note that the Linux kernel has opted
to do the same [2].

[1] - https://wiki.gentoo.org/wiki/Hardened/GNU_stack_quickstart
[2] - Linux kernel commit 0d362be5b142