Specify ELF program headers (PHDRS) in component.ld
This is not strictly necessary since Muen components are not loaded as
ELF files and the memory access rights are enforced as specified in the
component XML spec via the Tau0 generated pagetables. However, having
the PHDRS in the linker script fixes the following warning:

  ld: warning: sm has a LOAD segment with RWX permissions

Also move the .bss section after .data to have a more
traditional/familiar binary layout with the two data sections being
adjacent.