Set XSTATE_BV state-component bitmap in XSAVE header
According to Intel SDM Vol. 1, "13.8 Operation of XRSTOR", the `xrstor`
instruction uses a so called *request-feature bitmap* (RFBM) which is
defined as follows:

  RFBM := XCR0 and EDX:EAX;

The CPU updates each state-component *i* for which RFBM[i]=1.
Additionally to the RFBM, the value of XSTATE_BV[i] in the XSAVE area
header is considered as follows:

- If XSTATE_BV[i] = 0, the state component is set to its initial value.
- If XSTATE_BV[i] = 1, the state component is loaded with data from the
  XSAVE area.

Thus, in order for `xrstor` to consider the values set in the XSAVE
area, set the XSTATE_BV field to the active XCR0 features.