Rework detach key combo handling
Add separate state to handle partial match of detach key combo:

                     +------------+
                     | Processing |
                     +------------+
                     /             ^
           Attach   /               \  Detach/Match
                   v                 \
         +------------+    Match    +-----------+
         | Forwarding |------------>| Buffering |
         +------------+             +-----------+
               ^                          |
               |          Mismatch        |
               +--------------------------+

When the console is in forwarding mode and the first key of the detach
combinations is pressed, transition to the buffering state without
forwarding that key press immediately. After a second key press, if
there is a match again, perform the detach action and transition to the
processing state.
On the other hand, if there is a mismatch, forward the buffered as well
as the current data and transition back to the forwarding state.