Policy: Add console forwarding to more policies
Add the necessary config variable, channels, conditional mappings as
well as the necessary bootparams adjustments for Linux subjects to the
Integration, Integration KT as well as Mirage/Solo5 system policies.
The feature is disabled via the config variable
`dbgserver_console_forwarding` for Demo System VT-d as well as
Mirage/Solo5 while enabled for the integ systems.