Fix issue in Get_TSS_Descriptor procedure
Correct shift count of TSS_Address bits 31:24 from 56 to 32, see Intel
SDM Vol. 3A, 7.2.3 TSS Descriptor in 64-bit mode. This makes the
procedure work correctly for addresses having non-zero bits in 31:24.

This never was a problem until now because only addresses < 24 bit are
in use. The problem manifested in a new component with differing memory
layout.

Thanks to Ralf Hubert for spotting and fixing this issue.