Policy: Add extra xHCI RMRR region to DL380
The region must be accessible by usb_controller_1, otherwise the server
hangs on controller usage (dbgserver or Linux subjects). No ilo NMI gets
through, and the Muen SK is apparently not running in this state.

After some time (the exact time varies), the server continues running
normally.

The range has been identified by try-and-error from VT-d faults and is
incomplete. The current setting works well with demo_system_vtd and xHCI
assigned to dbgserver. If the server hangs with another VT-d fault in
the log, chances are that the xhci controller does not work correctly.
This needs to be investigated further (T678).