DM: Add support for Header_Type 16#80#
A header type value of 16#80# indicates a multi-function device (see PCI
specification 3.0, "6.2.1 Device Identification").

Allow this value in Emulate and mask out bit 7 in read emulation. This
announces the device as non-multi-function device to the DM client.

Note: We should probably announce multiple functions of the same
DM-managed device as multi-functional device. Will be implemented if
actually required.