emulate: Simplify QEMU cmd: drop ioh3420 I/O hub
The 3420 I/O hub (IOH) is not actually required, just use
bus=pcie.0,addr=2.0 for virtio-net-pci. This is in preparation of
updating coreboot/filo, which does not like this configuration (produces
a Linux kernel PF BUG when the virtio-net-pci driver is loaded, also
with "stock" Linux live CD, non-Muen that is).

Since we do not want to debug emulation-related bugs, just remove the
ioh3420 device from the QEMU command line. I checked, there is no
-device ioh3420 in VMs started by virt-manager/virsh.