Kernel/dbgserver: UART: Get rid of UART.State
Summary:
Update common submodule which adds the internal UART state to the
Put_Char procedure. Rationale: the inter-project package rename in the
kernel project does no longer work if common is built as a library.

For the kernel, hide internal state in SK.Console.[HSUART|UART] packages
in order to keep the console interface unchanged.

Test Plan:
Verified that building common as a library and using this library in the
kernel and dbgserver works.

Reviewers: ken

Reviewed By: ken

Differential Revision: https://dev.codelabs.ch/D1158