Move initial FPU state check to ITS component
The current initial FPU state assertions in the example subject rely on
the assumption that the FPU is not used before the component's
elaboration code. But the example component is under muinit control,
which also verifies hashes. Depending on the compiler/optimization
settings, the FPU might be used and the assertions fail.

Therefore move the FPU state check to the isolation test component,
which does not use muinit. This makes the FPU state check part of the
isolation test report.

Currently, only XCR0 and header fields are checked.