Add pciconf.py script
Generates the PCI device configuration for all devices assigned to a
subject. Also resolves device aliases and classes.

The device SID is still hard-coded, as it is unclear how to derive the
value at this stage (policy src). The SID is allocated by the expander
(if not specified explicitly at the subject-level).