Make Get_PCI_Capability more robust
If this procedure is called on a PCI configuration space without an
actual device, the Index is 16#ff#, which is not a valid capability
pointer, as it does not hold enough room for ID, next pointer and at
least one data byte.

Therefore, rename Capability_Range to Capability_Ptr_Type and reduce
upper limit by three bytes.