Initial per-device-mmconf operation
First shot at implementing an operation (Check_Vendor_Device) which
operates on a single device PCI configuration space. This is intended
for native drivers, which do not require access to the complete mmconf
array.

The aim is to re-use this code also in the procedures operating on the
complete mmconf array. This is achieved by relying on copy-by-reference
semantics when using the 'aliased' keyword as described by:

  http://www.ada-auth.org/standards/12rat/html/Rat12-4-2.html

This forces the compiler to pass non-tagged config space records by
reference. Note: access types to volatile objects (while supported in
SPARK) do not work as SPARK requires instances to be declared at library
level, which is not possible if the goal is high code reuse.

Approach verified working (build, proof and CI) using
- GNAT Pro 25.2
- GNAT FSF 12.2.0